refactor(config): extract alias migration
